Move allocator to separate files
