+++ dali/pipeline/operators/decoder/nvjpeg_decoder_cpu.h
@@ -25,6 +25,7 @@
 #include <vector>
 
 #include "dali/pipeline/operators/decoder/nvjpeg_helper.h"
+#include "dali/pipeline/operators/decoder/nvjpeg_allocator.h"
 
 #include "dali/image/image_factory.h"
 #include "dali/pipeline/operators/operator.h"
@@ -33,168 +34,6 @@
 
 namespace dali {
 
-namespace mem {
-// TODO(spanev): move in separate file
-class BasicPinnedAllocator {
- public:
-  static void PreallocateBuffers(size_t element_size_hint, size_t n_elements_hint) {
-    std::lock_guard<std::mutex> l(m_);
-
-    if (element_size_hint_ == 0) {
-      element_size_hint_ = element_size_hint;
-    } else if (element_size_hint_ != element_size_hint) {
-      DALI_FAIL("All instances of nvJPEGDecoder should have the same host_memory_padding.");
-    }
-    free_buffers_pool_.reserve(free_buffers_pool_.size() + n_elements_hint);
-    for (size_t i = 0; i < n_elements_hint; ++i) {
-      void* buffer;
-      CUDA_CALL(cudaHostAlloc(&buffer, element_size_hint, 0));
-      free_buffers_pool_.push_back(buffer);
-    }
-  }
-
-  static void FreeBuffers(size_t n_elements_hint) {
-    std::lock_guard<std::mutex> l(m_);
-    for (size_t i = 0; i < n_elements_hint; ++i) {
-      // Will call std::terminate
-      DALI_ENFORCE(free_buffers_pool_.size() > 0, "Some buffers were not returned by nvJPEG.");
-      CUDA_CALL(cudaFreeHost(free_buffers_pool_.back()));
-      free_buffers_pool_.pop_back();
-    }
-  }
-
-  static int Alloc(void** ptr, size_t size, unsigned int flags) {
-    std::lock_guard<std::mutex> l(m_);
-    // Non managed buffer
-    if (size > element_size_hint_ || free_buffers_pool_.empty()) {
-      return cudaHostAlloc(ptr, size, flags) == cudaSuccess ? 0 : 1;
-    }
-
-    // Managed buffer, adding it to the allocated set
-    void* buffer = free_buffers_pool_.back();
-    free_buffers_pool_.pop_back();
-    *ptr = buffer;
-    allocated_buffers_.insert(buffer);
-    return 0;
-  }
-
-  static int Free(void* ptr) {
-    std::lock_guard<std::mutex> l(m_);
-    if (allocated_buffers_.find(ptr) == allocated_buffers_.end()) {
-      // Non managed buffer... just free
-      return cudaFreeHost(ptr) == cudaSuccess ? 0 : 1;
-    }
-    // Managed buffer
-    allocated_buffers_.erase(ptr);
-    free_buffers_pool_.push_back(ptr);
-    return 0;
-  }
-
- private:
-  static std::vector<void*> free_buffers_pool_;
-  static size_t element_size_hint_;
-  static std::unordered_set<void*> allocated_buffers_;
-
-  static std::mutex m_;
-};
-
-class ChunkPinnedAllocator {
- public:
-  static void PreallocateBuffers(size_t element_size_hint, size_t n_elements_hint) {
-    std::lock_guard<std::mutex> l(m_);
-
-    counter_++;
-
-    if (element_size_hint_ == 0) {
-      element_size_hint_ = element_size_hint;
-    } else if (element_size_hint_ != element_size_hint) {
-      DALI_FAIL("All instances of nvJPEGDecoder should have the same host_memory_padding.");
-    }
-
-    Chunk chunk;
-    CUDA_CALL(cudaHostAlloc(&chunk.memory, element_size_hint * n_elements_hint, 0));
-    chunk.free_metadata.resize(n_elements_hint, true);
-    chunk.first_free = 0;
-    chunk.size = n_elements_hint;
-    chunks_.push_back(chunk);
-  }
-
-  static void FreeBuffers(size_t n_elements_hint) {
-    std::lock_guard<std::mutex> l(m_);
-    counter_--;
-    if (counter_ == 0) {
-      for (auto chunk : chunks_) {
-        // Will call std::terminate if CUDA_CALL fails
-        CUDA_CALL(cudaFreeHost(chunk.memory));
-      }
-      chunks_.clear();
-    }
-  }
-
-  static int Alloc(void** ptr, size_t size, unsigned int flags) {
-    std::lock_guard<std::mutex> l(m_);
-    // Non managed buffer
-    if (size > element_size_hint_) {
-      return cudaHostAlloc(ptr, size, flags) == cudaSuccess ? 0 : 1;
-    }
-
-    for (size_t c_idx = 0; c_idx < chunks_.size(); ++c_idx) {
-      auto& chunk = chunks_[c_idx];
-      size_t first_free = chunk.first_free;
-      // This chunk is full
-      if (chunk.first_free == chunk.size)
-        continue;
-
-      *ptr = static_cast<void*>(
-              static_cast<uint8_t*>(chunk.memory) + (element_size_hint_ * first_free));
-      allocated_buffers_[*ptr] = std::make_pair(c_idx, first_free);
-
-      // Update free metadata and find next first_free
-      chunk.free_metadata[first_free] = false;
-      while (first_free < chunk.size && !chunk.free_metadata[first_free]) {
-        first_free++;
-      }
-      chunk.first_free = first_free;
-      return 0;
-    }
-
-    return cudaHostAlloc(ptr, size, flags) == cudaSuccess ? 0 : 1;
-  }
-
-  static int Free(void* ptr) {
-    std::lock_guard<std::mutex> l(m_);
-    auto it = allocated_buffers_.find(ptr);
-    if (it == allocated_buffers_.end()) {
-      // Non managed buffer... just free
-      return cudaFreeHost(ptr) == cudaSuccess ? 0 : 1;
-    }
-    size_t c_idx, buff_idx;
-    std::tie(c_idx, buff_idx) = it->second;
-    auto& chunk = chunks_[c_idx];
-    chunk.free_metadata[buff_idx] = true;
-    if (chunk.first_free > buff_idx) {
-      chunk.first_free = buff_idx;
-    }
-    allocated_buffers_.erase(it);
-    return 0;
-  }
-
- private:
-  struct Chunk {
-    void* memory;
-    std::vector<int> free_metadata;
-    size_t first_free;
-    size_t size;
-  };
-  static std::vector<Chunk> chunks_;
-  static size_t element_size_hint_;
-  // ptr to chunk-offset pos
-  static std::unordered_map<void*, std::pair<size_t, size_t>> allocated_buffers_;
-
-  static int counter_;
-  static std::mutex m_;
-};
-}  // namespace mem
 
 using ImageInfo = EncodedImageInfo<unsigned int>;
